doc: fix bug with module relocation
